move ServiceTest from com.hjr.test to com.hjt.test.service
