+++ src/test/java/com/hjr/test/service/ClassServiceTest.java
@@ -1,4 +1,4 @@
-package com.hjr.test;
+package com.hjr.test.service;
 
 import com.hjr.been.Class;
 import com.hjr.service.ClassService;
