+++ src/test/java/com/hjr/test/service/StudentServiceTest.java
@@ -1,4 +1,4 @@
-package com.hjr.test;
+package com.hjr.test.service;
 
 import com.hjr.been.Student;
 import com.hjr.service.StudentService;
